Add separate polyfills file for loading shadow-dom
I'm thinking of using shadow-dom and customElements so that I can get CSS encapsulation of our components.

My strategy for polyfills is taken from here:

https://philipwalton.com/articles/loading-polyfills-only-when-needed/

Basically we have our app bundle, which contains no polyfills.  If the browser we're using has every feature we need, we just load the app bundle.

Then we have the polyfill bundle.  If the browser is missing even a single feature, and needs a single polyfill, we'll load the whole polyfill bundle.

This keeps the isolation clear on the development end, optimises for the up-to-date browser scenario, and isn't that bad if you have to load more polyfill code than is strictly required.