Add example of Custom Elements with CSS scoping.
Here we are rendering a custom element (with a macro to make the custom element class compatible with `customElements.define()`), and using ShadyCSS to scope the CSS so it doesn't escape.

See component.html for an example.

After all this, I think I won't go ahead with using ShadowDOM to scope the CSS, but rather use iFrames.  It's the extra security advantages that pushed me over the line, as well as realising I could use the `postMessage()` API to pass data back and forth with the main page.