Add more testing and a fix for Validators.
The fix was that "oneOf" previously failed on a null value, but should pass on a null value unless it is a required attribute.

I also clarified that "one of" uses strict equality, so should only be used with basic values.

Now all of the types are tested.
Still need to test required() and the extra utility functions.