Load client_enthralerdotcom.hxml through webpack
Use my `haxe_loader` library to load the hxml file.
This enables us to use `yarn run watch` for rebuilding, and `yarn run start` for both watching, rebuilding and refreshing the browser.  Not bad!

I had to move the webpack outputs into a separate "assets" folder for the dev server.