Just do a `cdnjs/` mapping rather than listing libraries one by one.
The only special case we'll do is jQuery - because it so common
and because of the weird way it interacts with AMD modules.

Key thing to note: anytime you're using a cdn import you do it like:

	'cdnjs/marked/0.3.6/marked.min'

Note that the ".js" is missing from the end - for some reason this
is needed with requirejs, if you include the ".js" on the end then
the remapping does not happen.