Get rid of EnthralerMessages.broadbastSchemaUrl and feature_shared_state feature flag
We don't need broadbastSchemaUrl on the main enthralerdotcom site, and I've updated EnthralerEditor.hx to just load the schema based on a URL parameter.

And I've removed the feature_shared_state stuff... if I do implement this, it'll be as an optional API, not a key part of the interface.